backend/requirements.txt
# Core Flask framework
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3

# Database & Supabase
supabase==1.2.0
psycopg2-binary==2.9.7

# Environment & Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Password hashing & Security
bcrypt==4.1.2
werkzeug==2.3.7

# Data processing & AI
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0

# File processing
PyPDF2==3.0.1
Pillow==10.0.0

# Date & Time utilities
python-dateutil==2.8.2

# API utilities
requests==2.31.0

# Validation
marshmallow==3.20.1

# Development utilities
python-json-logger==2.0.7